have all collections use IterableCollection
