+++ phpstan.neon
@@ -56,3 +56,7 @@ parameters:
             message: "#^Parameter \\#1 \\$value of function count expects array\\|Countable, iterable\\<K of \\(int\\|string\\), V\\> given\\.$#"
             count: 1
             path: src/Collection/LazyCollection.php
+        -
+            message: "#^Parameter \\#1 \\$p of method Doctrine\\\\Common\\\\Collections\\\\Collection\\<K of \\(int\\|string\\),V\\>\\:\\:filter\\(\\) expects Closure\\(V\\)\\: bool, Closure\\(V, K of \\(int\\|string\\)\\)\\: bool given\\.$#"
+            count: 1
+            path: src/Collection/DoctrineCollection.php
